Handle the case when a Bignum in JRuby contains a small value
In JRuby a Java extension can create a Bignum with any value, which our current implementation doesn't expect.

Ensure that the code handles byte arrays from Java's BigInteger#toByteArray that are less than 8 bytes – and make sure the top bytes are filled with either 0x00 or 0xFF depending on the sign of the value.

This fixes #112.