Avoid creating singleton classes when looking up extensions
`rb_singleton_class(VALUE)` creates the singleton class if it doesn't
exist yet, which isn't good for perf. First because it allocates
and object, but also because as far as I know it bust method caches.

Instead we can use `rb_class_of` which returns the singleton class
if the object has one, or the real class if it doesn't.

Later when we call `rb_obj_class`, we can compare the previous
class we got, if it's different it means the object had a singleton
class.