Optimize Packer and Unpacker initialize methods
Particularly through the common `MessagePack.load / dump`.

First we accept `nil` in place of the option hash, which allows
to skip several hash lookup when no option is passed.

Then if we actually pass an option hash, we used static
symbols instead of goign through the ID2SYM every time.

Finally for MessagePack.pack, we accept `nil` in place of
var args, as to avoid a costly `*rest` forwarding.