Implement lazy instantiation and Copy on Write for unpacker registry
Every time an `Unpacker` is instantiated we have to copy over a `VALUE[256]` so `1kiB`.
That's quite a lot of memory usage for such short lived objects, and copying isn't
that cheap.

So first we can only allocate that array when the first type is registered.
So in cases where no extension is used we simply don't incur the cost for it.

Then if we actually do need one, we can only copy it when it happens to be mutated.
In most cases, types are registered once during boot, but very rarely at runtime. So
in the vast majority of cases we should be able to just pass a pointer.

Every time we share a registry we increment it's `borrow_count`, every time a holder
is garbage collected, it decreate the counter, and the last one is in charge of freeing the registry.

Finally when registering a type in the registry, if it isn't yet shared, we
directly mutate it.

```ruby
require 'benchmark/ips'
require 'msgpack'

payload = MessagePack.dump([1, "2", 3.4])

Benchmark.ips do |x|
  x.report("load") { MessagePack.load(payload) }
  x.compare!
end
```

master:
```
$ ruby -Ilib bench/load.rb
Warming up --------------------------------------
                load    95.076k i/100ms
Calculating -------------------------------------
                load    931.850k (± 3.3%) i/s -      4.659M in   5.005056s
```

this branch:
```
Warming up --------------------------------------
                load   142.119k i/100ms
Calculating -------------------------------------
                load      1.442M (± 2.1%) i/s -      7.248M in   5.027389s
```