recursive extensions: mark the previous stacks
When we process a recursive extension, we replace the unpack stack by
a fresh one. The problem is that the stack contains VALUE references
so if GC triggers, we won't mark these objects and they'll bw garbage
collected.

To solve this `msgpack_unpacker_stack_t` now has a `parent` pointer,
and the unpacker mark function recursively scan the parent stacks.

Co-Authored-By: Peter Zhu <peter@peterzhu.ca>