Restore referenced buffers feature
I think this feature was accidentally removed in https://github.com/msgpack/msgpack-ruby/pull/213

On the MRI versions we support `rb_str_dup` won't copy the string
but return a shared string. So it's safe to always append a reference
when we are provided a string.

So instead of copying the RString content into the buffer, we simply
keep a reference to it, and point to its RSTRING_PTR().