Fix potential use-after-free in Buffer_free
`Buffer_free` rely on msgpack_buffer_t.owner to know wether
it should free the struct or if it's owned by another object.

This assumes that the "view" buffer will be freed before the object
that actually owns the buffer, but that's never guaranteed.

Instead we create view buffer without a `free` function, as they
have nothing to free anyway.

The only complexity is that when calling a method on a buffer, we
need to know if it's a direct buffer or a "view". To know this we
store that information in a regular instance variable.

Co-Authored-By: Peter Zhu <peter@peterzhu.ca>