Use rb_proc_call* rather than rb_funcall
It's a tiny bit faster, however it requires the receiver to be
an actual Proc.

By refactoring the register_type interface we can eagerly cast
all callbacks to actual procs.

In the end it makes the C side of the code simpler which is nice.

There is an extremely minor backward incompatibility in `#registered_types`
which no longer return the original callable, but always returns a Proc.