Implement a simpler regression test for #341
What is happening is that when we call into a recursive packing
proc, we first save the packer buffer state onto the stack
and then reset the buffer.

Once we return from the proc, the original buffer state is copied
back.

The problem with this is that if any of the chunk has a mapped string
then they are not reachable by any Ruby object and may be garbage collected
at any moment.