Mark buffers held onto the stack.
We instantiate a special Buffer that we keep on the stack
and is in change of marking the parent buffer we copied.

This really isn't ideal as we have to go over the linked list
to update the pointer to the tail.

The buffer implementation is in need of a more serious refactor
but this patch is a quick fix.