Packer: store the parent buffer mapped strings on the stack
Since the parent buffer isn't marked while we're executing the
extension proc, we need to ensure all the chunks `mapped_strings`
won't be GCed nor moved.

To do that, we copy all of them onto the stack.

It's a bit disgusting, but until we refactor the buffer implementation
to support recursive extensions better, it's the simplest solution.