Better fix for marking buffers held on the stack
Followup: https://github.com/msgpack/msgpack-ruby/pull/347

As discussed previously, the `alloca` hack was just a quick fix,
it's not good as it could lead to stack overflows and also could
be optimized away by compilers etc.

Here instead we copy all the VALUE references into a dedicated object
that is in change of holding and pinning these references.