Lazily allocate and eagerly free unpacker stack
Applications may have numerous unpacker instances allocated with
different configurations, but are very unlikely to have more than
one or a handful active concurrently.

As such eagerly allocating the 4kiB parsing stack is using more
memory than necessary.

Additionally, when only a few arena pages are in use, allocating
a chunk from a first page is very cheap.

The only complexity in this patch is the `Unpacker#each` API that
requires not to free the stack on EOF as parsing may be resumed
afterwards.